refactor title bar / command management code: paying that tech debt down!
