+++ src/packages/frontend/frame-editors/frame-tree/commands/button-bar.tsx
@@ -1,0 +1,3 @@
+export default function ButtonBar({}) {
+  return <div>TODO</div>;
+}
