+++ src/packages/frontend/frame-editors/frame-tree/commands/editor-menus.ts
@@ -101,7 +101,6 @@ export function addEditorMenus({
       );
     }
     if (!c.label) {
-      console.log("sub in ", name, c);
       c = { ...c, label: capitalize(name) };
     }
     const { children } = c;
