+++ src/packages/frontend/frame-editors/frame-tree/commands/index.ts
@@ -7,3 +7,4 @@ import "./generic-menus";
 import "./generic-commands";
 export { addEditorMenus } from "./editor-menus";
 import "./format-commands";
+export { ManageCommands } from "./manage";
